added the test_command function to srcs
